Finished implementing Flag.py suggestor
Fully implemented the script and its code properly, also ran --dev to submit changes, as well as added new feature to Get to allow you to exclude files from the search for list_files()
Updated .gitignore for the History created
Removed old tests
