+++ .gitignore
@@ -320,3 +320,5 @@ $RECYCLE.BIN/
 /CODE/SysInternal_Suite/.sys.ignore
 /ACCESS/
 /CODE/VulnScan/tools/NN features/
+/CODE/logicytics/User_History.json.gz
+/CODE/logicytics/User_History.json
