+++ CODE/config.ini
@@ -7,17 +7,39 @@ log_using_debug = false
 # Or would you like to append to the same log file?
 delete_old_logs = false
 # Logicytics will save preferences and history in a file,
-# This is used by Flag.py, to suggest better flags_list
+# This is used by Flag.py, to suggest better flags
 # Would you like this to happen?
+# This is recommended, as it will improve the suggestions - Data will never be shared
 save_preferences = true
 
 [System Settings]
 # Do not play with these settings unless you know what you are doing
-version = 3.2.1
-files = "bluetooth_details.py, bluetooth_logger.py, browser_miner.ps1, cmd_commands.py, config.ini, dir_list.py, dump_memory.py, event_log.py, Logicytics.py, log_miner.py, media_backup.py, netadapter.ps1, packet_sniffer.py, property_scraper.ps1, registry.py, sensitive_data_miner.py, ssh_miner.py, sys_internal.py, tasklist.py, tree.ps1, vulnscan.py, wifi_stealer.py, window_feature_miner.ps1, wmic.py, _debug.py, _dev.py, logicytics\Checks.py, logicytics\Execute.py, logicytics\FileManagement.py, logicytics\Flag.py, logicytics\Get.py, logicytics\Logger.py, logicytics\__init__.py, SysInternal_Suite\.sys.ignore, SysInternal_Suite\SysInternal_Suite.zip, VulnScan\Model SenseMini .3n3.pth, VulnScan\README.md, VulnScan\Vectorizer .3n3.pkl, VulnScan\tools\_study_network.py, VulnScan\tools\_test_gpu_acceleration.py, VulnScan\tools\_vectorizer.py, VulnScan\v2-deprecated\_generate_data.py, VulnScan\v3\_generate_data.py, VulnScan\v3\_train.py"
+version = 3.3.0
+files = "bluetooth_details.py, bluetooth_logger.py, browser_miner.ps1, cmd_commands.py, config.ini, dir_list.py, dump_memory.py, event_log.py, Logicytics.py, log_miner.py, media_backup.py, netadapter.ps1, packet_sniffer.py, property_scraper.ps1, registry.py, sensitive_data_miner.py, ssh_miner.py, sys_internal.py, tasklist.py, tree.ps1, vulnscan.py, wifi_stealer.py, window_feature_miner.ps1, wmic.py, _debug.py, _dev.py, logicytics\Checks.py, logicytics\Execute.py, logicytics\FileManagement.py, logicytics\Flag.py, logicytics\Get.py, logicytics\Logger.py, logicytics\User_History.json.gz, logicytics\__init__.py, SysInternal_Suite\.sys.ignore, SysInternal_Suite\SysInternal_Suite.zip, VulnScan\Model SenseMini .3n3.pth, VulnScan\README.md, VulnScan\Vectorizer .3n3.pkl, VulnScan\tools\_study_network.py, VulnScan\tools\_test_gpu_acceleration.py, VulnScan\tools\_vectorizer.py, VulnScan\v2-deprecated\_generate_data.py, VulnScan\v3\_generate_data.py, VulnScan\v3\_train.py"
 
 ###################################################
 # The following settings are for specific modules #
+###################################################
+
+[Flag Settings]
+# The minimum accuracy to suggest a flag,
+# This is a percentage, and must be a float
+# The default is 30.0, and is what we advise
+# If the accuracy is below this, the flag will move to the next suggestion process
+# The process is: difflib, then model, then history suggestions
+accuracy_min = 30.0
+
+# This is the model to use to suggest flags,
+# I advise to keep it as all-MiniLM-L6-v2
+# This is the best model for this task, and is lightweight
+# The model MUST be a Sentence Transformer model
+model_to_use = all-MiniLM-L6-v2
+
+# Finally, should debug mode be enabled for the flag module?
+# This will print out more information to the console,
+# This is for the model itself, and is based on tqdm, it shows extra info on batches
+model_debug = false
+
 ###################################################
 
 [PacketSniffer Settings]
@@ -105,3 +127,5 @@ vectorizer_path = PATH
 # Number of features to visualise in the SVG Bar graph, maximum is 3000 due to limitations
 # Placing -1 will visualise first 3000 features. Bar will be a color gradient heatmap.
 number_of_features = -1
+
+##################################################
